Change identifiers to `usize`
Change `Token::Identifier` to use `usize` instead of `String`, storing their identifiers in the state. Advantage is that less memory is consumed due to the sharing of strings and also less duplication occurs (most identifiers will be used more than once).

This creates a dependency of `Lexer` on `State`, which was inevitable.