Split into different mods
Split all the files into either `syntax`, `generate` or `runtime`. These represent the 3 stages of an Eko program. First, `syntax` is used to parse the raw source into an AST. Next, `generate` is used to convert that AST into bytecode. Finally, `runtime` is used to execute the code.

When Eko is used as a library, these mods should not be used at all. Instead, the convenience types `Engine` and `Compiler` should be used instead.