Change `InternalFuncReference` to store `Rc<Func>`
Change `InternalFuncReference` to store an `Rc<Func>` (which contains the function information) instead of cloning it. Store a shared version of the block is better, since the block might be rather large and cloning could cause a significant overhead.

In the future, if the language is switched to a garbage collector, blocks should still use `Rc`, as it has less overhead as is immutable anyway, which means no cycles.